% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyaudit-package.R
\docType{package}
\name{tidyaudit-package}
\alias{tidyaudit}
\alias{tidyaudit-package}
\title{tidyaudit: Pipeline Audit Trails and Data Diagnostics for 'tidyverse' Workflows}
\description{
Provides pipeline audit trails and data diagnostics for tidyverse workflows. The audit trail system captures lightweight metadata snapshots at each step of a pipeline, building a structured audit report without storing the data itself. Also includes diagnostic functions for interactive data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fpcordeiro/tidyaudit}
  \item Report bugs at \url{https://github.com/fpcordeiro/tidyaudit/issues}
}

}
\author{
\strong{Maintainer}: Fernando Cordeiro \email{fernandolpcordeiro@gmail.com} [copyright holder]

}
\keyword{internal}
